인터셉터 수정, 댓글 작성 폼 제작, 댓글 리스트 구현 중, 댓글 작성 구현

+++ src/main/java/com/kij/exam/demo/controller/UsrArticleController.java
@@ -47,7 +47,7 @@ public class UsrArticleController {
 
 		ResultData<Integer> writeArticleRd = articleService.writeArticle(rq.getLoginedMemberId(), boardId, title, body);
 
-		int id = (int) writeArticleRd.getData1();
+		int id = writeArticleRd.getData1();
 
 		return Utility.jsReplace(Utility.f("%d번 게시물을 작성했습니다", id), Utility.f("detail?id=%d", id));
 	}
@@ -173,5 +173,4 @@ public class UsrArticleController {
 
 		return "/usr/article/modify";
 	}
-
 }
